requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-telegram==0.19.0
requests==2.32.3
telegram_text==0.2.0
urllib3==2.2.3
